elgen-esg-model-service/models/requirements.txt
requests~=2.31.0
loguru==0.7.0
transformers==4.31.0
accelerate==0.20.3
pydantic~=1.10.9
torch==2.0.1
dependency-injector~=4.41.0
fastapi~=0.96.0
uvicorn~=0.22.0
einops~=0.6.1
numpy==1.25.1
scipy==1.11.1
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-sdk==1.18.0
python-json-logger==2.0.7
pytest~=7.4.0
pytest-cov~=4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
botocore~=1.29.101
httpx~=0.24.1
xformers~=0.0.20
bitsandbytes~=0.41.1
azure-storage-blob~=12.17.0
starlette~=0.27.0
safetensors~=0.3.2
huggingface-hub~=0.16.4